Refactor: Move Snowflake connection to shared
