+++ src/shared/snowflake/SnowflakeAuth.ts
